make concat dtypes the same for upgrade to tf2
(cherry picked from commit 26f520cbfb327f791b9388d9d689a499cfbe617a)
(cherry picked from commit 88ec5f55a7d969441d9cdc6cda12c8c891b345f6)
(cherry picked from commit 7a79f47cde0e111485260a8575d34eb6c0cc16c7)

+++ alf/utils/summary_utils.py
@@ -127,7 +127,8 @@ def histogram_continuous(name,
             bucket_counts = tf.cast(
                 tf.reduce_sum(input_tensor=one_hots, axis=0), dtype=tf.float64)
             edges = tf.linspace(bucket_min, bucket_max, bucket_count + 1)
-            edges = tf.concat([edges[:-1], [bucket_max]], 0)
+            edges = tf.concat([edges[:-1],
+                tf.cast([bucket_max], dtype=tf.float32)], 0)
             edges = tf.cast(edges, tf.float64)
             left_edges = edges[:-1]
             right_edges = edges[1:]
